adding test code

+++ tests/server_tests.py
@@ -261,7 +261,14 @@ def reset_data():
         Authorization.create(
             'haiti', 'reviewed_test_key',
             domain_write_permission='test.google.com',
-            mark_notes_reviewed=True,
+            mark_notes_reviewed=True),
+	Authorization.create(
+	    'haiti', 'not_allow_believed_dead_test_key',
+	    domain_write_permission='test.google.com',
+	    allow_believed_dead_permission=False),
+	Authorization.create(
+	    'haiti', 'allow_believed_dead_test_key',
+	    domain_write_permission='test.google.com',
 	    allow_believed_dead_permission=True),
         Authorization.create(
             '*', 'global_test_key',
@@ -709,7 +716,7 @@ class PersonNoteTests(TestsBase):
         error_message = page.first(class_=re.compile(r'.*\berror\b.*'))
         for fragment in fragments:
             assert fragment in error_message.text, (
-                '%s missing from error message' % fragment)
+                '%s missing from error message %s' % (fragment, error_message.text))
         self.s.back()
 
     # The verify_ functions below implement common fragments of the testing
@@ -846,6 +853,9 @@ class PersonNoteTests(TestsBase):
 
         # Do not assert params.  Upon reaching the details page, you've lost
         # the difference between seekers and providers and the param is gone.
+	#show_believed_dead_option = config.get_for_subdomain(subdomain,
+	#    'show_believed_dead_option')
+	#print "should be true %d " % show_believed_dead_option
         details_page = self.s.doc
         num_initial_notes = len(details_page.all(class_='view note'))
         note_form = details_page.first('form')
@@ -859,7 +869,7 @@ class PersonNoteTests(TestsBase):
             params['status'] = status
             extra_values.append(str(NOTE_STATUS_TEXT.get(status)))
 
-        details_page = self.s.submit(note_form, **params)
+	details_page = self.s.submit(note_form, **params)
         notes = details_page.all(class_='view note')
         assert len(notes) == num_initial_notes + 1
         new_note_text = notes[-1].text
@@ -1052,6 +1062,8 @@ class PersonNoteTests(TestsBase):
 
         # Start on the home page and click the "I'm looking for someone" button
         self.go('/?subdomain=haiti')
+	show_believed_dead_option = config.get_for_subdomain('haiti',
+		'show_believed_dead_option')
         search_page = self.s.follow('I\'m looking for someone')
         search_form = search_page.first('form')
         assert 'Search for this person' in search_form.content
@@ -1110,28 +1122,28 @@ class PersonNoteTests(TestsBase):
         entry.delete()
 
         # Add a note with status == 'believed_dead'.
-        self.verify_update_notes(
-            True, '_test Third note body', '_test Third note author',
-            'believed_dead')
+	if show_believed_dead_option:
+            self.verify_update_notes(
+                True, '_test Third note body', '_test Third note author',
+                'believed_dead')
+            # Check that a UserActionLog entry was created.
+            entry = UserActionLog.all().get()
+            assert entry.action == 'mark_dead'
+            assert entry.detail == '_test_first_name _test_last_name'
+            assert entry.ip_address
+            assert entry.Note_text == '_test Third note body'
+            assert entry.Note_status == 'believed_dead'
+            entry.delete()
 
-        # Check that a UserActionLog entry was created.
-        entry = UserActionLog.all().get()
-        assert entry.action == 'mark_dead'
-        assert entry.detail == '_test_first_name _test_last_name'
-        assert entry.ip_address
-        assert entry.Note_text == '_test Third note body'
-        assert entry.Note_status == 'believed_dead'
-        entry.delete()
+            person = Person.all().filter('first_name =', '_test_first_name').get()
+            assert person.entry_date == datetime.datetime(2006, 6, 6, 6, 6, 6)
 
-        person = Person.all().filter('first_name =', '_test_first_name').get()
-        assert person.entry_date == datetime.datetime(2006, 6, 6, 6, 6, 6)
-
-        self.s.submit(search_form, query='_test_first_name')
-        assert_params()
-        self.verify_results_page(
-            1, all_have=['_test_first_name'], some_have=['_test_first_name'],
-            status=['Someone has received information that this person is dead']
-        )
+            self.s.submit(search_form, query='_test_first_name')
+            assert_params()
+            self.verify_results_page(
+                1, all_have=['_test_first_name'], some_have=['_test_first_name'],
+                status=['Someone has received information that this person is dead']
+            )
 
         # test for default_expiry_days config:
         config.set_for_subdomain('haiti', default_expiry_days=10)
@@ -1293,6 +1305,7 @@ class PersonNoteTests(TestsBase):
         # Try a multiword match on an alternate name.
         self.s.submit(search_form, query='ABCD EFG QRST UVWX')
         self.verify_results_page(1, all_have=(['ABCD EFGH']))
+
 
     def test_indexing_japanese_names(self):
         """Index Japanese person's names and make sure they are searchable."""
@@ -1741,6 +1754,10 @@ class PersonNoteTests(TestsBase):
         """Test the posting and viewing of the note status field in the UI."""
         status_class = re.compile(r'\bstatus\b')
 
+	# show_believed_dead_option = True
+	config.set_for_subdomain('haiti', show_believed_dead_option=True)
+	show_dead = config.get_for_subdomain('haiti', 'show_believed_dead_option')
+ 
         # Check that the right status options appear on the create page.
         doc = self.go('/create?subdomain=haiti&role=provide')
         note = doc.first(class_='note input')
@@ -1798,6 +1815,81 @@ class PersonNoteTests(TestsBase):
 
         # Check that a UserActionLog entry was not created.
         assert not UserActionLog.all().get()
+
+	# show_believed_dead_option = False
+	config.set_for_subdomain('japan', show_believed_dead_option=False)
+	show_dead = config.get_for_subdomain('japan', 'show_believed_dead_option')
+
+	# Check that believed_dead option does not appear on the create page
+	doc = self.go('/create?subdomain=japan&role=provide')
+	note = doc.first(class_='note input')
+	options = note.first('select', name='status').all('option')
+	assert len(options) == len(NOTE_STATUS_OPTIONS) - 1
+	for option, text in zip(options, NOTE_STATUS_OPTIONS):
+	    assert text in option.attrs['value']
+	    assert option.attrs['value'] != 'believed_dead'
+
+        # Create a record with no status and get the new record's ID.
+        form = doc.first('form')
+        doc = self.s.submit(form,
+                            first_name='_test_first',
+                            last_name='_test_last',
+                            author_name='_test_author',
+                            text='_test_text')
+        view_url = self.s.url
+
+        # Check that the right status options appear on the view page.
+        doc = self.s.go(view_url)
+        note = doc.first(class_='note input')
+        options = note.first('select', name='status').all('option')
+	assert len(options) == len(NOTE_STATUS_OPTIONS) - 1
+	for option, text in zip(options, NOTE_STATUS_OPTIONS):
+	    assert text in option.attrs['value']
+	    assert option.attrs['value'] != 'believed_dead'
+
+        # Check that the right status options appear on the view page.
+        doc = self.s.go(view_url)
+        note = doc.first(class_='note input')
+        options = note.first('select', name='status').all('option')
+        assert len(options) == len(NOTE_STATUS_OPTIONS) - 1
+        for option, text in zip(options, NOTE_STATUS_OPTIONS):
+            assert text in option.attrs['value']
+	    assert option.attrs['value'] != 'believed_dead'
+
+        # Set the status in a note and check that it appears on the view page.
+        form = doc.first('form')
+        self.s.submit(form, author_name='_test_author2', text='_test_text',
+                                    status='believed_alive')
+        doc = self.s.go(view_url)
+        note = doc.last(class_='view note')
+        assert 'believed_alive' in note.content
+        assert 'believed_dead' not in note.content
+
+        # Check that a UserActionLog entry was created.
+        entry = UserActionLog.all().get()
+        assert entry.action == 'mark_alive'
+        assert entry.detail == '_test_first _test_last'
+        assert not entry.ip_address
+        assert entry.Note_text == '_test_text'
+        assert entry.Note_status == 'believed_alive'
+        entry.delete()
+
+        # Check that a UserActionLog entry was created.
+	# Set status to believed_dead, but show_believed_dead_option is false.
+	self.s.submit(form,
+		      author_name='_test_author',
+		      text='_believed_dead_test_text',
+		      status='believed_dead')
+	self.assert_error_deadend(
+	    self.s.submit(form,
+			  author_name='_test_author',
+			  text='_test_text',
+			  status='believed_dead'),
+	    'permission', 'Status of this person')
+
+	# Check that a UserActionLog entry was not created.
+	assert not UserActionLog.all().get()
+
 
     def test_api_write_pfif_1_2(self):
         """Post a single entry as PFIF 1.2 using the upload API."""
@@ -2099,6 +2191,38 @@ class PersonNoteTests(TestsBase):
         # Confirm all notes are marked reviewed.
         for note in notes:
             assert note.reviewed == True
+
+    def test_api_allow_believed_dead_permission(self):
+	""" Test whether the API key is authorized to report a person dead. """
+	# Add the associated person record to the datastore
+        data = get_test_data('test.pfif-1.2.xml')
+        self.go('/api/write?subdomain=haiti&key=test_key', 
+	        data=data, type='application/xml')
+
+	# Test authorized key.
+	data = get_test_data('test.pfif-1.2-believed-dead.xml')
+	doc = self.go('/api/write?subdomain=haiti&key=allow_believed_dead_test_key',
+	        data=data, type='application/xml')
+	person = Person.get('haiti', 'test.google.com/person.21009')
+	notes = person.get_notes()
+	# Confirm the newly-added note with status believed_dead
+	for note in notes:
+	    if note.note_record_id == 'test.google.com/note.218':
+		assert note.status == 'believed_dead'
+
+	# Test unauthorized key.
+	doc = self.go('/api/write?subdomain=haiti&key=not_allow_believed_dead_test_key',
+                data=data, type='application/xml')
+	# The Person record should not be updated
+	person_status = doc.first('status:write')
+	assert person_status.first('status:written').text == '0'	
+	# The Note record should be rejected with error message
+	note_status = person_status.next('status:write')
+	assert note_status.first('status:parsed').text == '1'
+	assert note_status.first('status:written').text == '0'
+	assert ('Not authorized to mark person with dead status' in
+		note_status.first('status:error').text)
+	
 
     def test_api_subscribe_unsubscribe(self):
         """Subscribe and unsubscribe to e-mail updates for a person via API"""
@@ -4477,6 +4601,7 @@ class PersonNoteTests(TestsBase):
         assert '_test_last' not in first_title
         person.delete()
 
+
     def test_config_family_name_first(self):
         # family_name_first=True
         doc = self.go('/create?subdomain=china')
@@ -4648,6 +4773,31 @@ class PersonNoteTests(TestsBase):
         assert '_test_alternate_first' not in first_title
         assert '_test_alternate_last' not in first_title
         person.delete()
+
+
+    def test_config_show_believed_dead_option(self):
+	# show_believed_dead_option=True
+	config.set_for_subdomain('haiti', show_believed_dead_option=True)
+	doc = self.go('/create?subdomain=haiti')
+	self.s.submit(doc.first('form'),
+                      first_name='_test_first',
+                      last_name='_test_last',
+                      author_name='_test_author')
+        person = Person.all().get()
+        doc = self.go('/view?id=%s&subdomain=haiti' % person.record_id)
+ 	assert doc.all('option', value='believed_dead')
+
+	# show_believed_dead_option=False
+	config.set_for_subdomain('japan', show_believed_dead_option=False)
+	doc = self.go('/create?subdomain=japan')
+	self.s.submit(doc.first('form'),
+                      first_name='_test_first',
+                      last_name='_test_last',
+                      author_name='_test_author')
+        person = Person.all().get()
+        doc = self.go('/view?id=%s&subdomain=japan' % person.record_id)
+ 	assert not doc.all('option', value='believed_dead')
+
 
     def test_config_use_postal_code(self):
         # use_postal_code=True
@@ -4891,6 +5041,7 @@ class ConfigTests(TestsBase):
             family_name_first='false',
             use_alternate_names='false',
             use_postal_code='false',
+	    show_believed_dead_option='false',
             min_query_word_length='1',
             map_default_zoom='6',
             map_default_center='[4, 5]',
@@ -4910,6 +5061,7 @@ class ConfigTests(TestsBase):
         assert not cfg.family_name_first
         assert not cfg.use_alternate_names
         assert not cfg.use_postal_code
+	assert not cfg.show_believed_dead_option
         assert cfg.min_query_word_length == 1
         assert cfg.map_default_zoom == 6
         assert cfg.map_default_center == [4, 5]
@@ -4926,6 +5078,7 @@ class ConfigTests(TestsBase):
             family_name_first='true',
             use_alternate_names='true',
             use_postal_code='true',
+	    show_believed_dead_option='true',
             min_query_word_length='2',
             map_default_zoom='7',
             map_default_center='[-3, -7]',
@@ -4945,6 +5098,7 @@ class ConfigTests(TestsBase):
         assert cfg.family_name_first
         assert cfg.use_alternate_names
         assert cfg.use_postal_code
+	assert cfg.show_believed_dead_option
         assert cfg.min_query_word_length == 2
         assert cfg.map_default_zoom == 7
         assert cfg.map_default_center == [-3, -7]
